fix: rename file with created name

+++ readlineHandlers/fileOperations/rn.js
@@ -9,11 +9,12 @@ export const rn = async (input, directory) => {
     if (params.length !== 2) throw new Error("Invalid input");
 
     const lastFilePath = path.resolve(directory["currentDirectory"], params[0]);
-    const isFile = await checkPath(lastFilePath, "file");
-    if (!isFile) throw new Error("Invalid path to file");
-
     const lastFileDirectory = path.dirname(lastFilePath);
     const newFilePath = path.join(lastFileDirectory, params[1]);
+    const isFile = await checkPath(lastFilePath, "file");
+    const isNewFile = await checkPath(newFilePath, "file");
+    if (!isFile || isNewFile) throw new Error("Invalid path to file");
+
     await fs.rename(lastFilePath, newFilePath);
   } catch (err) {
     let errMessage = "Operation failed";
